app/requirements.txt
fastapi>=0.95
uvicorn
sqlalchemy
pydantic>=2.0
pydantic-settings>=2.0
passlib>=1.7,<1.8
bcrypt>=3.1,<3.3
python-jose[cryptography]
python-multipart
python-dotenv
alembic
pymysql

pytest
pytest-asyncio
httpx
